share project details fetch function
