+++ ui/pages/Project/Project.jsx
@@ -4,9 +4,10 @@ import { connect } from 'react-redux'
 import { Route, Switch } from 'react-router-dom'
 import { Loader } from 'semantic-ui-react'
 
+import { getProjectDetailsIsLoading } from 'redux/selectors'
 import { Error404 } from 'shared/components/page/Errors'
 import { loadCurrentProject, unloadProject } from './reducers'
-import { getProjectDetailsIsLoading, getCurrentProject } from './selectors'
+import { getCurrentProject } from './selectors'
 import ProjectPageUI from './components/ProjectPageUI'
 import CaseReview from './components/CaseReview'
 import FamilyPage from './components/FamilyPage'
