+++ ui/pages/Project/components/ProjectPageUI.jsx
@@ -4,6 +4,7 @@ import { connect } from 'react-redux'
 import { Grid, Loader } from 'semantic-ui-react'
 import { Link } from 'react-router-dom'
 
+import { getProjectDetailsIsLoading } from 'redux/selectors'
 import DataLoader from 'shared/components/DataLoader'
 import { HorizontalSpacer, VerticalSpacer } from 'shared/components/Spacers'
 import { SectionHeader } from 'shared/components/StyledComponents'
@@ -16,7 +17,6 @@ import {
 } from 'shared/utils/constants'
 import {
   getCurrentProject,
-  getProjectDetailsIsLoading,
   getAnalysisStatusCounts,
   getProjectOverviewIsLoading,
   getFamiliesLoading,
